# openapi.yaml
openapi: 3.1.0
info:
  title: PinkHive Paid Utility API
  version: 0.2.0
  description: Deterministic website-prep, guard, receipt, and salary-disclosure products paid through x402.
servers:
  - url: https://pinkhive-website-prep-x402.onrender.com
paths:
  /health:
    get:
      operationId: health
      responses:
        "200":
          description: Service and payment readiness
  /v1/pricing:
    get:
      operationId: pricing
      responses:
        "200":
          description: Current billable-unit definition
  /v1/site-prep:
    post:
      operationId: buildSitePrepBundle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteRequest"
      responses:
        "200":
          description: Generated bundle; one successful metered unit
        "400":
          description: Invalid request, not metered
        "401":
          description: Invalid prepaid key
        "402":
          description: Payment required
  /v1/claude-ship-guard:
    post:
      operationId: buildClaudeShipGuard
      summary: Generate and deliver one customized Claude Code commit-guard ZIP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShipGuardRequest"
      responses:
        "200":
          description: Base64-encoded deterministic ZIP with SHA-256 and file inventory
        "400":
          description: Invalid customization request
        "402":
          description: Base USDC payment required through x402
  /v1/salary-disclosure-audit:
    post:
      operationId: auditSalaryDisclosure
      summary: Audit salary disclosure fields in JSON-LD JobPosting objects
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema: {type: string, minLength: 1, maxLength: 255}
          description: Retrying with the same key returns the original metering receipt without another meter unit.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/SalaryAuditRequest"}
      responses:
        "200": {description: One metered audit request}
        "400": {description: Invalid body, not metered}
        "402": {description: Base USDC payment required through x402}
components:
  schemas:
    Page:
      type: object
      required: [url, html]
      properties:
        url: {type: string, format: uri}
        html: {type: string}
    ShipGuardRequest:
      type: object
      required: [project_name]
      properties:
        project_name: {type: string, minLength: 1, maxLength: 80}
        commit_style: {type: string, enum: [descriptive, conventional], default: descriptive}
        secret_presets:
          type: array
          uniqueItems: true
          minItems: 1
          maxItems: 3
          items: {type: string, enum: [cloud, github, generic]}
        max_subject_length: {type: integer, minimum: 50, maximum: 100, default: 72}
    SalaryAuditRequest:
      type: object
      additionalProperties: false
      required: [job_postings]
      properties:
        job_postings:
          type: array
          minItems: 1
          maxItems: 25
          items:
            type: object
            required: ["@type"]
            properties:
              "@type":
                oneOf:
                  - const: JobPosting
                  - type: array
                    contains: {const: JobPosting}
    SiteRequest:
      type: object
      required: [site_name, site_summary, pages]
      properties:
        site_name: {type: string}
        site_summary: {type: string}
        pages:
          type: array
          minItems: 1
          maxItems: 20
          items: {$ref: "#/components/schemas/Page"}